moved fn to utils
