+++ components/ui/answerOptions.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { shuffleArray } from "@/lib/utils";
 import React, { useMemo } from "react";
 
 interface Option {
@@ -15,14 +16,6 @@ interface AnswerOptionsProps {
     onSelectionChange: (value: number | number[]) => void;
 }
 
-const shuffleArray = (array: Option[]): Option[] => {
-    const copy = [...array];
-    for (let i = copy.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [copy[i], copy[j]] = [copy[j], copy[i]];
-    }
-    return copy;
-};
 
 const AnswerOptions: React.FC<AnswerOptionsProps> = ({
     options,
